+++ next.config.ts
@@ -2,7 +2,6 @@ import type { NextConfig } from "next";
 
 const nextConfig: NextConfig = {
   output: "export",
-  basePath: process.env.NODE_ENV === "production" ? "/personal-website" : "",
   allowedDevOrigins: ["10.0.0.237"],
   experimental: {
     viewTransition: true,